fix: delegation withdrawal while validator pending

+++ builtin/staker/staker.go
@@ -441,11 +441,13 @@ func (s *Staker) WithdrawStake(validator thor.Address, endorser thor.Address, cu
 		if err != nil {
 			return 0, err
 		}
-		if err := s.globalStatsService.AddWithdrawable(exit.QueuedDecrease); err != nil {
-			return 0, err
-		}
-		if err := s.globalStatsService.RemoveQueued(exit.QueuedDecrease); err != nil {
-			return 0, err
+		if exit.QueuedDecrease > 0 {
+			if err := s.globalStatsService.AddWithdrawable(exit.QueuedDecrease); err != nil {
+				return 0, err
+			}
+			if err := s.globalStatsService.RemoveQueued(exit.QueuedDecrease); err != nil {
+				return 0, err
+			}
 		}
 	}
